requirements.txt
asgiref==3.3.4
brotlipy==0.7.0
certifi==2020.12.5
cffi @ file:///C:/ci/cffi_1606237165371/work
chardet @ file:///C:/ci/chardet_1605296577955/work
click==8.0.0
colorama==0.4.4
conda==4.10.1
conda-package-handling @ file:///C:/ci/conda-package-handling_1605484860147/work
cryptography @ file:///C:/ci/cryptography_1607618721363/work
-e git+https://github.com/django/django.git@530f58caaa5052e9e56bf8461caee4d821953bcb#egg=Django
Flask==2.0.0
idna @ file:///tmp/build/80754af9/idna_1593446292537/work
itsdangerous==2.0.0
Jinja2==3.0.0
joblib @ file:///tmp/build/80754af9/joblib_1613502643832/work
manual-spellchecker==1.2
MarkupSafe==2.0.0
menuinst==1.4.16
mkl-fft==1.3.0
mkl-random @ file:///C:/ci/mkl_random_1618854182866/work
mkl-service==2.3.0
nltk==3.6.2
numpy==1.20.3
pandas==1.2.4
pycosat==0.6.3
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
pyenchant==3.2.0
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1606517880428/work
PySocks @ file:///C:/ci/pysocks_1605307512533/work
python-aiml==0.9.3
python-dateutil==2.8.1
pytz==2021.1
pywin32==228
regex==2021.4.4
requests @ file:///tmp/build/80754af9/requests_1606691187061/work
ruamel-yaml-conda @ file:///C:/ci/ruamel_yaml_1605524207026/work
scikit-learn @ file:///C:/ci/scikit-learn_1614446712466/work
scipy @ file:///C:/ci/scipy_1618856124610/work
six @ file:///C:/ci/six_1605205404068/work
sqlparse==0.4.1
threadpoolctl @ file:///tmp/tmp9twdgx9k/threadpoolctl-2.1.0-py3-none-any.whl
tqdm @ file:///tmp/build/80754af9/tqdm_1607369919789/work
urllib3 @ file:///tmp/build/80754af9/urllib3_1603305693037/work
Werkzeug==2.0.0
win-inet-pton @ file:///C:/ci/win_inet_pton_1605306162074/work
wincertstore==0.2
